feat(image): register multichannel images
This commit adds a function to register multichannel images using
pystackreg and a transformation matrix created from a single reference
channel.

cellstream is a package designed to extract data from multichannel
fluorescence microscopy images. I find that sometimes my images (whether
through technique or equipment or random chance) are jittery and
require image registration to be properly handled by the image
processing functions. I felt that adding a function to handle what I
assume to be a common issue would fit well with the spirit of this
project.

As a bonus, the function is designed to gracefully handle both tensors
and numpy arrays as input. `register_multichannel()` will return a
registered image of the same type that was passed to it.